+++ packages/yarnpkg-core/sources/LightReport.ts
@@ -2,7 +2,7 @@ import {Writable}                from 'stream';
 
 import {Configuration}           from './Configuration';
 import {MessageName}             from './MessageName';
-import {Report}                  from './Report';
+import {Report, TimerOptions}    from './Report';
 import {formatNameWithHyperlink} from './StreamReport';
 import * as formatUtils          from './formatUtils';
 import {Locator}                 from './types';
@@ -58,12 +58,18 @@ export class LightReport extends Report {
   reportCacheMiss(locator: Locator) {
   }
 
-  startTimerSync<T>(what: string, cb: () => T) {
-    return cb();
+  startTimerSync<T>(what: string, opts: TimerOptions, cb: () => T): void;
+  startTimerSync<T>(what: string, cb: () => T): void;
+  startTimerSync<T>(what: string, opts: TimerOptions | (() => T), cb?: () => T) {
+    const realCb = typeof opts === `function` ? opts : cb!;
+    return realCb();
   }
 
-  async startTimerPromise<T>(what: string, cb: () => Promise<T>) {
-    return await cb();
+  async startTimerPromise<T>(what: string, opts: TimerOptions, cb: () => Promise<T>): Promise<void>;
+  async startTimerPromise<T>(what: string, cb: () => Promise<T>): Promise<void>;
+  async startTimerPromise<T>(what: string, opts: TimerOptions | (() => Promise<T>), cb?: () => Promise<T>) {
+    const realCb = typeof opts === `function` ? opts : cb!;
+    return await realCb();
   }
 
   async startCacheReport<T>(cb: () => Promise<T>) {